requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
dynaconf==3.1.8
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.0.1
greenlet==3.0.3
groq==0.5.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isapi-wsgi==0.4.2
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
psycopg2==2.9.4
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dotenv==0.19.2
sniffio==1.3.1
SQLAlchemy==2.0.29
toml==0.10.2
typing_extensions==4.11.0
waitress==3.0.0
Werkzeug==3.0.2